refactor: moved cryptography modules into folder
